[Feat] Add edit event component
addressed comments
