+++ client/src/components/Event/Event.scss
@@ -39,7 +39,7 @@
   .event {
     &__buttons {
       display: inline;
-     
+
       & * {
         margin-top: 5px;
       }
@@ -54,7 +54,7 @@
     &__buttons {
       padding: 0 2%;
       display: inline;
-     
+
       & * {
         margin-top: 5px;
       }
