+++ client/src/components/EventUpdateForm/EventUpdateForm.scss
@@ -44,11 +44,13 @@
     flex-direction: column;
   }
 }
+
 @media only screen and (max-width: 768px) {
   .form {
     width: 60vw;
   }
 }
+
 @media only screen and (max-width: 600px) {
   .form {
     width: 90vw;
